Quantities are now vertically aligned in the cart. Next up: refactor checkout sheet to be its own file. ContentView is way too long.
