+++ Tappayo/common/CurrencyTextField.swift
@@ -41,7 +41,11 @@ struct CurrencyTextField: UIViewRepresentable {
         textField.placeholder = placeholder
         textField.keyboardType = .numberPad
         textField.delegate = context.coordinator
-        textField.textAlignment = .left // Left align text
+        textField.textAlignment = .left
+
+        // Apply rounded border style to match SwiftUI's .roundedBorder
+        textField.borderStyle = .roundedRect
+        textField.backgroundColor = UIColor.systemGray6
 
         // Convert SwiftUI Font to UIFont and apply
         let uiFont = UIFont.preferredFont(from: font)
